Update pal from commit b59afda2
* Adding y216 y210 y416 y410
* [CTS_Next] dEQP-VK.memory.device_memory_report.* failures
* Pad PS_IN up to VS_OUT+1 on Gfx11.
* Use colorful error msgs with gcc/clang
* Enable ExecuteIndirect Packet support on APUs
* Fix IB2 info size
* Implemented app detect for watchdogs legions scratch
* fix bug in MemoryCacheLayer
* Allow disabling ACE via panel setting
* remove unused RMT URI service
* gfx11VertexAttributesRingBufferSizePerSe key description
* Fix Noisy Asserts & Alerts
* Optimize MSAA copies on "Z" modes.
* Initialize DescriptorTable base VirtAddr in a corner case.
* Allow specification of read/write UMC counters
* Remove unused Crash Analysis struct
* Slightly optimize basic shaders.
* Do not override IFH PAL mode when spoofing GPU
* Add dynamic state vertexBufferCount
* Linux Implicit Gang-Submit
* Call OnConfigUpdated in UpdateTraceConfig
* Reset CmdBufferId when entering Begin
* 64-bit palArchiveFile
* Initialize CmdBuffer MemoryChunk CPU-side
* Correlation information does not come for some binds
* Navi3X set the RLC_SPM_RING_WRPTR during the SPM_Begin.
* [Navi3x] Set DisablePremablePipelineStats by default so some HW interfaces can be clock gated
* Remove Chip specific detects
* Add MCBP support for Navi3x on Linux
* Support compute queue with Gfx9::HwlFastColorClear.
* [Navi3x][CTS] dEQP-VK.query_pool.statistics_query.compute_* tests failing when EnablePreamblePipelineStats is disabled

+++ src/util/fileArchiveCacheLayer.h
@@ -80,9 +80,9 @@ private:
     // Container types
     struct Entry
     {
-        uint32 ordinalId;
-        uint32 dataSize;
-        uint32 storeSize;
+        uint64 ordinalId;
+        uint64 dataSize;
+        uint64 storeSize;
     };
     using EntryMap = HashMap<EntryKey,
                              Entry,
@@ -100,7 +100,7 @@ private:
     Result RefreshHeaders();
 
     // Helper function for constructor
-    static uint32 GetHashMapNumBuckets(const IArchiveFile* pArchiveFile);
+    static size_t GetHashMapNumBuckets(const IArchiveFile* pArchiveFile);
 
     // Invariants that must be passed in by ctor
     IArchiveFile* const  m_pArchivefile;
